moved disaggregated over to fairness
