+++ credoai/evaluators/fairness.py
@@ -7,7 +7,7 @@ from credoai.evaluators.utils.validation import (
     check_data_instance,
     check_existence,
 )
-from credoai.evidence import MetricContainer
+from credoai.evidence import MetricContainer, TableContainer
 from credoai.modules.metric_constants import MODEL_METRIC_CATEGORIES
 from credoai.modules.metrics import Metric, find_metrics
 from credoai.utils import global_logger
@@ -56,11 +56,11 @@ class ModelFairness(Evaluator):
         self.fairness_metrics = None
         self.fairness_prob_metrics = None
 
-    name = "Fairness"
+    name = "ModelFairness"
     required_artifacts = {"model", "data", "sensitive_feature"}
 
     def _setup(self):
-        self.sensitive_features = self.data.sensitive_feature.name
+        self.sensitive_features = self.data.sensitive_feature
         self.y_true = self.data.y
         self.y_pred = self.model.predict(self.data.X)
         if hasattr(self.model, "predict_proba"):
@@ -85,14 +85,24 @@ class ModelFairness(Evaluator):
         fairness_results = self.get_fairness_results()
         fairness_results = pd.DataFrame(fairness_results).reset_index()
         fairness_results.rename({"metric_type": "type"}, axis=1, inplace=True)
-        label = {"sensitive_feature": fairness_results.sensitive_feature.iloc[0]}
+        disaggregated_df = self.get_disaggregated_performance()
+
         self.results = [
             MetricContainer(
                 fairness_results.drop("sensitive_feature", axis=1),
                 **self.get_container_info(
                     labels={"sensitive_feature": self.sensitive_features.name}
                 ),
-            )
+            ),
+            TableContainer(
+                disaggregated_df,
+                **self.get_container_info(
+                    labels={
+                        "sensitive_feature": self.sensitive_features.name,
+                        "metric_types": disaggregated_df.type.unique().tolist(),
+                    }
+                ),
+            ),
         ]
         return self
 
@@ -127,6 +137,33 @@ class ModelFairness(Evaluator):
             self.y_true,
             self.sensitive_features,
         )
+
+    def get_disaggregated_performance(self):
+        """Return performance metrics for each group
+
+        Parameters
+        ----------
+        melt : bool, optional
+            If True, return a long-form dataframe, by default False
+
+        Returns
+        -------
+        pandas.DataFrame
+            The disaggregated performance metrics
+        """
+        disaggregated_df = pd.DataFrame()
+        for metric_frame in self.metric_frames.values():
+            df = metric_frame.by_group.copy().convert_dtypes()
+            disaggregated_df = pd.concat([disaggregated_df, df], axis=1)
+        disaggregated_results = disaggregated_df.reset_index().melt(
+            id_vars=[disaggregated_df.index.name],
+            var_name="type",
+        )
+        disaggregated_results.name = "disaggregated_performance"
+
+        if disaggregated_results.empty:
+            global_logger.warn("No disaggregated metrics could be calculated.")
+        return disaggregated_results
 
     def get_fairness_results(self):
         """Return fairness and performance parity metrics
